backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.1
tensorflow==2.16.1
numpy==1.26.4
pillow==11.2.1
python-multipart==0.0.20
